+++ CHANGELOG.md
@@ -38,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `polygonizeLines()` from `PGS_Processing`, in favour of `polygonize(PShape)`.
 * `unionMeshWithoutHoles()` from `PGS_ShapeBoolean`. Previously deprecated in favour of the more general `unionMesh()`.
 * `fromGeoJSON()` and `toGeoJSON()` from `PGS_Conversion`.
+* The `COARSE` mesh coloring algorithm, since it can color adjacent faces the same colour.
 
 ## **2.1** *(2025-10-04)*